\name{as.character.parental.list}
\alias{as.character.parental.list}
\title{Convert a 'parental...}
\usage{as.character.parental.list(x, pretty=F, ...)}
\description{Convert a 'parental.list' to a character vector.}
\value{A character vector.}
\arguments{\item{x}{A object of class 'parental'.}
\item{pretty}{A logical of length 1 indicating if the character vectors 
are pretty-printed.}
\item{...}{Further arguments (unused)}}
