\name{as.cpdag.bn}
\alias{as.cpdag.bn}
\title{Convert a parental BN to a...}
\usage{\method{as.cpdag}{bn} (x, ...)}
\description{Convert a parental BN to a 
Completed Partially Directed Acyclic Graph (CPDAG).}
\value{The CPDAG as a parental. (Which will in general not be a BN)}
\arguments{\item{x}{The parental BN to be converted.}
\item{...}{Further arguments (unused)}}
