\name{bvs}
\alias{bvs}
\title{Constructor function for a 'bvs' object.}
\usage{bvs(...)}

\description{
  Constructor function for a 'bvs' object.
}

\details{
  'bvs' is a subclass of \code{parental}. See
  \code{\link{parental}} for detailed documentation.
}
\value{An object of class 'bvs'.}
\arguments{
  \item{...}{A series of vectors specifying the parents of each node. 
These vectors must be of storage.mode "integer".}
}
