\name{bvsresponse.list}
\alias{bvsresponse.list}
\title{Constructor function for a 'bvsresponse list' object.}
\usage{
  bvsresponse.list(...)
}
\arguments{
  \item{...}{A series of objects of class 'bvsresponse'.}
}
\value{
  An object of class 'bvsresponse.list'
}
\description{
  Constructor function for a 'bvsresponse list' object.
}

