\name{complete}
\alias{complete}
\title{Return a complete graph.}
\usage{
  complete(n)
}
\arguments{
  \item{n}{the number of nodes}
}
\value{
  A complete \code{parental} graph
}
\description{
  Return a complete graph.
}
\examples{
complete(5)
}
\seealso{
  \code{\link{empty}}, \code{\link{parental}}
}

