\name{convertToEnlargedCoordinates}
\alias{convertToEnlargedCoordinates}
\title{Expand coordinates.}
\usage{
  convertToEnlargedCoordinates(width, height, x, y)
}
\arguments{
  \item{width}{A vector of widths?}

  \item{height}{A vector of heights?}

  \item{x}{x-coordinates}

  \item{y}{y-coordinates}
}
\value{
  A list with two components: \code{width} and
  \code{height}.
}
\description{
  Expand the coordinates. Not currently used
}
\seealso{
  \code{\link{grplot}}
}

