\name{coordinatesBetweenRectangles}
\alias{coordinatesBetweenRectangles}
\title{.}
\usage{coordinatesBetweenRectangles(x1, y1, x2, y2, w1, h1, w2, h2)}
\description{...}
\details{...}
\value{...}
\arguments{\item{x1}{...}
\item{y1}{...}
\item{x2}{...}
\item{y2}{...}
\item{w1}{...}
\item{h1}{...}
\item{w2}{...}
\item{h2}{...}}
