\name{cptBinary}
\alias{cptBinary}
\title{Create a binary Conditional Probability Table...}
\usage{cptBinary(i, j)}
\description{Create a binary Conditional Probability Table}
\seealso{\code{\link{simulate.bn}}}
\arguments{\item{i}{The probability of value 1}
\item{j}{The probability of value 2}}
\examples{cptBinary(0.2, 0.8)}
