\name{distance}
\alias{distance}
\title{Calculate distance between two points.}
\usage{
  distance(x, y)
}
\arguments{
  \item{x}{A numeric vector of length 2, containing the
  x-coordinates of the two points}

  \item{y}{A numeric vector of length 2, containing the
  y-coordinates of the two points}
}
\value{
  A numeric vector of length 1. The distance.
}
\description{
  A vectorised function for calculating the distance
  between two points in Cartesian space. Computes the
  distance between \code{(x[1], y[1])} and \code{(x[2],
  y[2])}.
}

