\name{getbvsrl}
\alias{"[.bvsresponse.list"}
\alias{getbvsrl}
\title{Extract parts of a 'bvsresponse list'.}
\usage{
  \method{[}{bvsresponse.list}(x, i)
}
\arguments{
  \item{x}{Object from which to extract element(s)}

  \item{i}{Indicies specifying elements to extract}
}
\value{
  An object of class \code{bvsresponse.list}, consisting
  the parts of \code{x} indicated by \code{i}
}
\description{
  Operator that acts on objects of class
  \code{bvsresponse.list} to extract parts of the list
}

