\name{is.valid}
\alias{is.valid}
\title{Check validity.}
\usage{
  is.valid(x)
}
\arguments{
  \item{x}{Object to check for validity}
}
\value{
  A logical indicating validity.
}
\description{
  Checks the validity of objects
}
\seealso{
  \code{\link{is.valid.parental}},
  \code{\link{is.valid.bn}}, \code{\link{is.valid.bvs}},
  \code{\link{is.valid.bvsresponse}}
}

