\name{latticeGrob}
\alias{latticeGrob}
\title{Convert a lattice object to a grob.}
\usage{
  latticeGrob(p, ...)
}
\arguments{
  \item{p}{A lattice object}

  \item{...}{further arguments}
}
\value{
  A grob
}
\description{
  See Paul Murrell's message to r-help - Mar 21, 2010
  8:21:39 pm. "Re: [R] lattice grob"
  http://markmail.org/message/tg6kdxsr74fynncy
}

