\name{marginal.probs}
\alias{marginal.probs}
\title{Undocumented.}
\usage{
  marginal.probs(bn, ptables, N)
}
\arguments{
  \item{bn}{...}

  \item{ptables}{...}

  \item{N}{...}
}
\description{
  I don't think this works. In particular check that
  sample.int is sensible. And that N is properly
  implemented
}

