\name{maximallyOrientEdges.matrix}
\alias{maximallyOrientEdges.matrix}
\title{Undocumented.}
\usage{\method{maximallyOrientEdges}{matrix} (pdag, verbose=F,
    ...)}

\description{
  Undocumented.
}

\details{
  ...
}
\value{...}
\arguments{
  \item{pdag}{...}
  \item{verbose}{...}
  \item{...}{further arguments (unused)}
}
