\name{maximallyOrientEdges.parental}
\alias{maximallyOrientEdges.parental}
\title{Undocumented.}
\usage{maximallyOrientEdges.parental(pdag, verbose=F, ...)}
\description{Undocumented.}
\details{...}
\value{...}
\arguments{\item{pdag}{...}
\item{verbose}{A logical}
\item{...}{Further arguments (unused)}}
