\name{maximallyOrientEdgesSlow.matrix}
\alias{maximallyOrientEdgesSlow.matrix}
\title{Maximally orient edges.}
\usage{
  maximallyOrientEdgesSlow.matrix(pdag, verbose = F, ...)
}
\arguments{
  \item{pdag}{A pdag}

  \item{verbose}{A logical of length 1.}

  \item{...}{Further arguments (unused)}
}
\value{
  Something?
}
\description{
  This is the R implementation
}

