\name{nodeLevelplot}
\alias{nodeLevelplot}
\title{Multi-page plot of nodeLevelPlots.}
\usage{
  nodeLevelplot(parents, rawdata)
}
\arguments{
  \item{parents}{An object of class 'parental'}

  \item{rawdata}{A data frame.}
}
\description{
  Creates a plot with a level plot for each variable, each
  of which is presented on a separate page.
}

