\name{panel.parental}
\alias{panel.parental}
\title{xxxx...}
\usage{panel.parental(x, y, parents, layout, col, alpha, edgecol,
    edgealpha, islist=F, rawdata, grobNode, grobNodeSize,
    offset, widthMultiplier=2, heightMultipler=1, ...)}
\description{xxxx}
\details{....}
\value{....}
\arguments{\item{x}{....}
\item{pretty}{...}}
