\name{paste.and.draw2}
\alias{paste.and.draw2}
\title{function title...}
\usage{paste.and.draw2(left, right, sep=" : ", horizontal=TRUE, center=TRUE,
    showl=TRUE, showr=TRUE, gp=lattice:::gpar())}
\description{function title}
\details{description}
\arguments{\item{left}{...}
\item{right}{...}
\item{sep}{...}
\item{horizontal}{...}
\item{center}{...}
\item{showl}{...}
\item{showr}{...}
\item{gp}{...}}
