\name{pdag2alldagsSlow.matrix}
\alias{pdag2alldagsSlow.matrix}
\title{Undocumented.}
\usage{\method{pdag2alldagsSlow}{matrix} (pdag, verbose=F, ...)}

\description{
  Undocumented.
}

\details{
  ....
}
\value{....}
\arguments{
  \item{pdag}{....}
  \item{verbose}{...}
  \item{...}{Further arguments (unused)}
}
