\name{pdag2alldagsSlow.parental}
\alias{pdag2alldagsSlow.parental}
\title{Undocumented.}
\usage{
  \method{pdag2alldagsSlow}{parental} (pdag, verbose = F,
  ...)
}
\arguments{
  \item{pdag}{A pdag}

  \item{verbose}{...}

  \item{...}{Further arguments (unused)}
}
\value{
  ....
}
\description{
  ....
}

