\name{pltabulate}
\alias{pltabulate}
\title{Tabulate a parental list...}
\usage{pltabulate(pl, pretty=F, levels, sort=F)}
\description{Tabulate a parental list}
\details{Tabulate a parental list.}
\arguments{\item{pl}{A \code{parental.list} to tabulate.}
\item{pretty}{A logical indicating how to convert the objects of
class \code{parental} inside \code{pl} to strings. See 
\code{\link{as.character.parental.list}} for details.}
\item{levels}{What levels to make the table using. By default 
\code{as.character(pl)}.}
\item{sort}{A logical indicating whether to sort the table.}}
