\name{prepanel.parental}
\alias{prepanel.parental}
\title{xxxx...}
\usage{prepanel.parental(x, y, parents, rawdata, grobNodeSize, offset, islist=F,
    widthMultiplier=2, heightMultipler=1)
}
\description{xxxx}
\details{....}
\value{....}
\arguments{\item{x}{....}
\item{y}{...}
\item{parents}{...}
\item{rawdata}{...}
\item{grobNodeSize}{...}
\item{offset}{...}
\item{islist}{...}
\item{widthMultiplier}{...}
\item{heightMultipler}{...}
}

