\name{print.bn}
\alias{print.bn}
\title{Prints 'bn'.}
\usage{
  \method{print}{bn} (x, ...)
}
\arguments{
  \item{x}{A 'bn' object}

  \item{...}{Further arguments (unused)}
}
\value{
  Prints the 'bn' object to the console.
}
\description{
  Prints 'bn'.
}
\examples{
x <- bn(c(), 1, 2)
x
}
\seealso{
  \code{\link{bn}}
}

