\name{print.bvsresponse}
\alias{print.bvsresponse}
\title{Prints 'bvsresponse'.}
\usage{\method{print}{bvsresponse} (x, ...)}
\description{Prints 'bvsresponse'.}
\details{Prints a 'bvsresponse' object to the console.}
\value{Prints the 'bvsresponse' object to the console.}
\arguments{\item{x}{A 'bvsresponse' object}
\item{...}{Further arguments, currently unused}}
