\name{print.parental.list}
\alias{print.parental.list}
\title{Prints a 'parental...}
\usage{\method{print}{parental.list} (x, ...)}
\description{Prints a 'parental.list' object to the console.}
\value{Prints the 'parental.list' object to the console.}
\arguments{\item{x}{A 'parental.list' object}
\item{...}{Further arguments (unused)}}
