\name{recurse_unoriented_edge}
\alias{recurse_unoriented_edge}
\title{Recurse unoriented edge.}
\usage{
  recurse_unoriented_edge(cpdag, dag_list, verbose = F)
}
\arguments{
  \item{cpdag}{A cpdag}

  \item{dag_list}{...}

  \item{verbose}{A logical indicating whether verbose
  output should be given.}
}
\value{
  Unknown
}
\description{
  ...
}

