\name{recurse_unoriented_edgeSlow}
\alias{recurse_unoriented_edgeSlow}
\title{Recurse unoriented edge (slow).}
\usage{
  recurse_unoriented_edgeSlow(cpdag, dag_list, verbose = F)
}
\arguments{
  \item{cpdag}{...}

  \item{dag_list}{...}

  \item{verbose}{...}
}
\value{
  ...
}
\description{
  This is the R implementation
}

