\name{renameNodes.parental.list}
\alias{renameNodes.parental.list}
\title{Rename nodes.}
\usage{\method{renameNodes}{parental.list} (x, newnames, ...)}

\description{
  Rename nodes.
}

\details{
  Return the parental list, with nodes in each parental in
  the parental.list renamed to the newnames.
}
\value{The parental.list with renamed nodes}
\arguments{
  \item{x}{A parental.list}
  \item{newnames}{A character vector specifying the new names for the 
nodes}
  \item{...}{Further arguments, currently unused}
}
