\name{simulate.bn}
\alias{simulate.bn}
\title{.}
\usage{simulate.bn(object, nsim, seed, ptables, expectation=F, ...)}
\description{...}
\details{....}
\arguments{\item{object}{A bn}
\item{nsim}{Number of simulations}
\item{seed}{The starting seed (UNIMPLEMENTED)}
\item{ptables}{...}
\item{expectation}{...}
\item{...}{Further arguments (unused)}}
