\name{strip.default2}
\alias{strip.default2}
\title{Undocumented.}
\usage{strip.default2(which.given, which.panel, var.name, factor.levels,
    shingle.intervals, strip.names=c(FALSE, TRUE), strip.levels=c(TRUE,
    FALSE), sep=" : ", style=1, horizontal=TRUE,
    bg=trellis.par.get("strip.background")$col[which.given],
    fg=trellis.par.get("strip.shingle")$col[which.given],
    par.strip.text=trellis.par.get("add.text"), dimx)}
\description{Undocumented.}
\details{description}
\arguments{\item{which.given}{...}
\item{which.panel}{...}
\item{var.name}{...}
\item{factor.levels}{...}
\item{shingle.intervals}{...}
\item{strip.names}{...}
\item{strip.levels}{...}
\item{sep}{...}
\item{style}{...}
\item{horizontal}{...}
\item{bg}{...}
\item{fg}{...}
\item{par.strip.text}{...}
\item{dimx}{ADDED ...}}
