\name{toCartesian}
\alias{toCartesian}
\title{Convert polar coordinates to Cartesian...}
\usage{toCartesian(r, theta)}
\description{Convert polar coordinates to Cartesian}
\details{Currently unused. Not sure if this is vectorised}
\value{A list with two components, named \code{x} and \code{y}, 
containing the coordinates of the supplied point.}
\arguments{\item{r}{Polar coordinates}
\item{theta}{Polar coordinates}}
