\name{toPolar}
\alias{toPolar}
\title{Convert cartesian coordinates to polars.}
\usage{
  toPolar(x, y)
}
\arguments{
  \item{x}{Cartesian x-coordinates}

  \item{y}{Cartesian y-coordinates}
}
\value{
  A list with two components, named \code{r} and
  \code{theta}, containing the radius and angle to the
  supplied point.
}
\description{
  Currently unused. Not sure if this is vectorised
}

