\name{add.undirected.edge}
\alias{add.undirected.edge}
\title{Add undirected edge...}
\usage{add.undirected.edge(x, from, to, check.cycles, update=T, debug=F)}
\description{Add undirected edge}
\details{....}
\arguments{\item{x}{...}
\item{from}{...}
\item{to}{...}
\item{check.cycles}{...}
\item{update}{...}
\item{debug}{...}}
