\name{as.bn.graphNEL}
\alias{as.bn.graphNEL}
\title{Convert a graphNEL object to a 'bn' object.}
\usage{as.bn.graphNEL(x, ...)}
\description{Convert a graphNEL object to a 'bn' object. graphNEL objects are 
from bioconductor package 'graph'}
\details{Note that graphNEL's edge list are *children* lists.}
\value{An object of class 'bn'.}
\arguments{\item{x}{An object of class 'graphNEL'}
\item{...}{unused}}
