\name{as.cpdag0.bn.list}
\alias{as.cpdag0.bn.list}
\title{Convert 'bn list' to CPDAG (old version).}
\usage{
  \method{as.cpdag0}{bn.list} (x, ...)
}
\arguments{
  \item{x}{An object of class bn.list}

  \item{...}{Further arguments (unused)}
}
\value{
  A parental.list containing a list of CPDAGs of class
  CPDAG.
}
\description{
  Convert a bn.list to list of parental.list of Completed
  Partially Directed Acyclic Graph (CPDAG).
}

