\name{as.graph.parental}
\alias{as.graph.parental}
\title{Convert a 'parental' object to a 'graphNEL' object.}
\usage{as.graph.parental(x, ...)
}
\description{Convert a 'parental' object to a 'graphNEL' object. graphNEL objects are 
from bioconductor package 'graph'}
\details{Note that graphNEL's edge list are *children* lists.}
\value{An object of class 'graphNEL'.}
\arguments{\item{x}{An object of class 'parental'}
\item{...}{unused}
}

