\name{as.parental.bn}
\alias{as.parental.bn}
\title{Convert a 'bn' to a 'parental'...}
\usage{\method{as.parental}{bn} (x, ...)}
\description{Convert a 'bn' to a 'parental'}
\value{The object 'x', but of class 'parental'}
\arguments{\item{x}{An object of class 'bn'}
\item{...}{Further arguments (unused)}}
\examples{x <- bn(c(), c(1), c(1, 2))
as.parental(x)}
