\name{as.parental.character}
\alias{as.parental.character}
\title{Convert a character string to a parental...}
\usage{\method{as.parental}{character} (x, pretty=F, ...)}
\description{Convert a character string to a parental}
\details{Format required is one of two. Either standard or pretty.}
\value{....}
\arguments{\item{x}{An object}
\item{pretty}{A logical indicating whether the supplied object is pretty}
\item{...}{Further arguments (unused)}}
