\name{as.parental.graphNEL}
\alias{as.parental.graphNEL}
\title{Convert a graphNEL object to a 'parental' object.}
\usage{as.parental.graphNEL(x)
}
\description{Convert a graphNEL object to a 'parental' object. graphNEL objects are 
from bioconductor package 'graph'}
\details{Note that graphNEL's edge list are *children* lists.}
\value{An object of class 'parental'.}
\arguments{\item{x}{An object of class 'graphNEL'}
}

