\name{barchart.parental}
\alias{barchart.parental}
\title{Barchart of a graph...}
\usage{barchart.parental(x, data, type=c("indegrees", "outdegrees"), ...)}
\description{Barchart of a graph}
\details{Plot a barchart of summary statistics about a 'parental' object.}
\value{A barchart}
\arguments{\item{x}{An object of class parental}
\item{data}{currently unused}
\item{type}{A character vector of length one. One of \code{indegree}, 
\code{outdegree}}
\item{...}{Further arguments, passed onto \code{barchart()}}}
