\name{bn.list}
\alias{bn.list}
\title{Create a bn...}
\usage{bn.list(...)
bn.list(...)
}
\description{Create a bn.list}
\details{\code{bn.list}: Creates a list of objects of \code{bn} class.

\code{bn.list}: Creates a list of objects of \code{bn} class.


}
\value{\code{bn.list}: An object of class \code{bn.list}, containing all supplied BNs.

\code{bn.list}: An object of class \code{bn.list}, containing all supplied BNs.


}
\arguments{\item{...}{A number of objects of class \code{bn}}
}
\alias{bn.list}

