\name{c.parental.list}
\alias{c.parental.list}
\title{Concatenates a 'parental...}
\usage{\method{c}{parental.list} (...)}
\description{Concatenates a 'parental.list' object to the console.}
\value{An new 'parental.list' object, including all the supplied 
parental.lists}
\arguments{\item{...}{Any number of 'parental.list' objects}}
