\name{checkAcyclic}
\alias{checkAcyclic}
\title{Acyclicity testing...}
\usage{checkAcyclic(parental)}
\description{Acyclicity testing}
\details{Check for cycles in a directed graph.}
\value{A logical. TRUE if \code{parental} is acyclic. FALSE if 
\code{parental} contains a cycle.}
\arguments{\item{parental}{An object of class \code{parental}}}
