\name{convertToEnlargedCoordinates}
\alias{convertToEnlargedCoordinates}
\title{Expand coordinates.}
\usage{convertToEnlargedCoordinates(width, height, x, y)}
\description{Expand coordinates.}
\details{Expand the coordinates. Not currently used}
\value{A list with two components: \code{width} and \code{height}.}
\seealso{\code{\link{grplot}}}
\arguments{\item{width}{A vector of widths?}
\item{height}{A vector of heights?}
\item{x}{x-coordinates}
\item{y}{y-coordinates}}
