\name{drawDetails.lattice}
\alias{drawDetails.lattice}
\title{drawDetails for lattice plots.}
\usage{
  drawDetails.lattice(x, recording = F)
}
\arguments{
  \item{x}{....}

  \item{recording}{...}
}
\value{
  ....
}
\description{
  See Paul Murrell's message to r-help - Mar 21, 2010
  8:21:39 pm. "Re: [R] lattice grob"
  http://markmail.org/message/tg6kdxsr74fynncy
}

