\name{empty}
\alias{empty}
\title{Returns an empty graph with n nodes, of the given class...}
\usage{empty(n, class="parental", response)}
\description{Returns an empty graph with n nodes, of the given class}
\value{An empty graph of the specified class.}
\arguments{\item{n}{A integer of length 1, specifying the number of nodes}
\item{class}{The class of the returned graph. Can be any of "parental", "bn" or "bvs".}}
