\name{filterCyclic}
\alias{filterCyclic}
\title{Filter cyclic graphs...}
\usage{filterCyclic(bnlist)
}
\description{Filter cyclic graphs}
\details{Removes the cyclic graphs from a list of graphs}
\value{An object of class \code{bn.list}, containing all the acyclic graphs 
in \code{bnlist}}
\arguments{\item{bnlist}{A \code{parental.list} of graphs}
}

