\name{getbnl}
\alias{getbnl}
\alias{getbnl}
\title{Extract parts of a bn...}
\usage{[.bn.list(x, i)
}
\description{Extract parts of a bn.list}
\details{Operator that acts on objects of class \code{bn.list} to extract 
parts of the list}
\value{An object of class \code{bn.list}, consisting the parts of 
\code{x} indicated by \code{i}}
\arguments{\item{x}{Object from which to extract element(s)}
\item{i}{Indicies specifying elements to extract}
}

