\name{getbvsrl}
\alias{getbvsrl}
\alias{getbvsrl}
\title{Extract parts of a bvsresponse...}
\usage{[.bvsresponse.list(x, i)
}
\description{Extract parts of a bvsresponse.list}
\details{Operator that acts on objects of class \code{bvsresponse.list} to extract 
parts of the list}
\value{An object of class \code{bvsresponse.list}, consisting the parts of 
\code{x} indicated by \code{i}}
\arguments{\item{x}{Object from which to extract element(s)}
\item{i}{Indicies specifying elements to extract}
}

