\name{getpl}
\alias{getpl}
\title{Extract parts of a parental...}
\usage{[.parental.list(x, i)}
\description{Extract parts of a parental.list}
\details{Operator that acts on objects of class \code{parental.list} to extract 
parts of the list}
\value{An object of class \code{parental.list}, consisting the parts of 
\code{x} indicated by \code{i}}
\arguments{\item{x}{Object from which to extract element(s)}
\item{i}{Indicies specifying elements to extract}}
