\name{grobNodeNameSize}
\alias{grobNodeNameSize}
\title{Return the size of node "node" when the name of the node is plotted...}
\usage{grobNodeNameSize(node, parents, rawdata)
}
\description{Return the size of node "node" when the name of the node is plotted}
\value{A list of length 2 containing two items:
width: A object of class "unit"
height: A object of class "unit"}
\arguments{\item{node}{An integer of length 1, indicating which node the dimensions
should be computed for.}
\item{parents}{An object of class "parental" containing the graph that is to 
be plotted}
\item{rawdata}{The rawdata}
}

