\name{grplot.bvsresponse}
\alias{grplot.bvsresponse}
\title{Plot a 'bvsresponse' graph.}
\usage{grplot.bvsresponse(x, col="default", ...)}
\description{Plot a 'bvsresponse' graph.}
\value{A lattice plot of the graph}
\arguments{\item{x}{A 'bvsresponse' object}
\item{col}{A vector with each component indicating the colour of the 
corresponding node. The default character vector vector "default"
makes the response node red and the other nodes black.}
\item{...}{further arguments}}
