\name{grplot.parental.list}
\alias{grplot.parental.list}
\title{xxxx...}
\usage{grplot.parental.list(parentallist, col=1, alpha=1, edgecol=1,
    edgealpha=1, layout.par=list(niter = 2e+05),
    grobNode=grobNodeName, grobNodeSize=grobNodeNameSize,
    offset=0.25, ...)}
\description{xxxx}
\details{....}
\value{....}
\arguments{\item{x}{....}
\item{pretty}{...}}
