\name{grplot.parental}
\alias{grplot.parental}
\title{xxxx...}
\usage{grplot.parental(parents, col=1, alpha=1, edgecol, edgealpha=1,
    layout.par=list(niter = 1e+05), grobNode=grobNodeName,
    grobNodeSize=grobNodeNameSize, offset=0.25, layout, ...)}
\description{xxxx}
\details{....}
\value{....}
\arguments{\item{parents}{....}
\item{col}{...}
\item{alpha}{...}
\item{edgecol}{...}
\item{edgealpha}{...}
\item{layout.par}{...}
\item{grobNode}{...}
\item{grobNodeSize}{...}
\item{offset}{...}
\item{layout}{...}
\item{...}{...}}
