\name{indegrees.parental}
\alias{indegrees.parental}
\title{Indegrees...}
\usage{indegrees.parental(x, ...)}
\description{Indegrees}
\details{Get the indegree of each node of the supplied graph}
\value{A vector of length \code{nNodes(x)}, with the indegrees of each 
node of \code{x}}
\arguments{\item{x}{An object of class \code{parental}}
\item{...}{Further arguments (unused)}}
