\name{is.valid.parental}
\alias{is.valid.parental}
\title{Checks whether the supplied parental is valid.}
\usage{is.valid.parental(x)}
\description{Checks whether the supplied \code{parental} is valid.
Tests that the parents are sorted correctly, are of 
storage.mode() == "integer"}
\value{A logical of length 1 indicating whether x is a valid 
\code{parental} object}
\arguments{\item{x}{A object of class \code{parental}}}
