\name{is.valid}
\alias{is.valid}
\title{Check validity.}
\usage{is.valid(x)}
\description{Check validity.}
\details{Checks the validity of objects}
\value{A logical indicating validity.}
\seealso{\code{\link{is.valid.parental}},
\code{\link{is.valid.bn}}, \code{\link{is.valid.bvs}},
\code{\link{is.valid.bvsresponse}}}
\arguments{\item{x}{Object to check for validity}}
