\name{marginal.probs}
\alias{marginal.probs}
\title{Undocumented.}
\usage{marginal.probs(bn, ptables, N)}

\description{
  Undocumented.
}

\details{
  I don't think this works. In particular check that
  sample.int is sensible. And that N is properly
  implemented
}
\arguments{
  \item{bn}{...}
  \item{ptables}{...}
  \item{N}{...}
}
