\name{maximallyOrientEdges.matrix}
\alias{maximallyOrientEdges.matrix}
\title{.}
\usage{maximallyOrientEdges.matrix(pdag, verbose=F)
maximallyOrientEdges.matrix(pdag, verbose=F)
}
\description{...}
\details{\code{maximallyOrientEdges.matrix}: ...

\code{maximallyOrientEdges.matrix}: ...


}
\value{\code{maximallyOrientEdges.matrix}: ...

\code{maximallyOrientEdges.matrix}: ...


}
\arguments{\item{pdag}{...}
\item{verbose}{...}
}
\alias{maximallyOrientEdges.matrix}

