\name{maximallyOrientEdges.matrix}
\alias{maximallyOrientEdges.matrix}
\title{Undocumented.}
\usage{
  \method{maximallyOrientEdges}{matrix} (pdag, verbose = F,
  ...)
}
\arguments{
  \item{pdag}{...}

  \item{verbose}{...}

  \item{...}{further arguments (unused)}
}
\value{
  ...
}
\description{
  ...
}

