\name{maximallyOrientEdgesSlow.matrix}
\alias{maximallyOrientEdgesSlow.matrix}
\title{Maximally orient edges.}
\usage{maximallyOrientEdgesSlow.matrix(pdag, verbose=F, ...)}

\description{
  Maximally orient edges.
}

\details{
  This is the R implementation
}
\value{Something?}
\arguments{
  \item{pdag}{A pdag}
  \item{verbose}{A logical of length 1.}
  \item{...}{Further arguments (unused)}
}
