\name{nNodes.parental}
\alias{nNodes.parental}
\title{Number of nodes/vertices...}
\usage{nNodes.parental(parental)}
\description{Number of nodes/vertices}
\details{Get the number of nodes in a \code{parental} object}
\value{The number of nodes in \code{parental}, an integer.}
\arguments{\item{parental}{An object of class \code{parental}}}
