\name{panel.parental}
\alias{panel.parental}
\title{xxxx...}
\usage{panel.parental(x, y, parents, layout, col, alpha, edgecol, edgealpha, islist=F,
    rawdata, grobNode, grobNodeSize, offset, widthMultiplier=2,
    heightMultipler=1, ...)
}
\description{xxxx}
\details{....}
\value{....}
\arguments{\item{x}{....}
\item{y}{...}
\item{parents}{...}
\item{layout}{...}
\item{col}{...}
\item{alpha}{...}
\item{edgecol}{...}
\item{edgealpha}{...}
\item{islist}{...}
\item{rawdata}{...}
\item{grobNode}{...}
\item{grobNodeSize}{...}
\item{offset}{...}
\item{widthMultiplier}{...}
\item{heightMultipler}{...}
\item{...}{...}
}

