\name{print.bn}
\alias{print.bn}
\title{Prints a 'bn' object to the console.}
\usage{\method{print}{bn} (x, ...)}
\description{Prints a 'bn' object to the console.}
\value{Prints the 'bn' object to the console.}
\seealso{\code{\link{bn}}}
\arguments{\item{x}{A 'bn' object}
\item{...}{Further arguments (unused)}}
\examples{x <- bn(c(), 1, 2)
x}
