\name{print.bvsresponse}
\alias{print.bvsresponse}
\title{Prints 'bvsresponse'.}
\usage{
  \method{print}{bvsresponse} (x, ...)
}
\arguments{
  \item{x}{A 'bvsresponse' object}

  \item{...}{Further arguments, currently unused}
}
\value{
  Prints the 'bvsresponse' object to the console.
}
\description{
  Prints a 'bvsresponse' object to the console.
}

