\name{print.bvsresponse}
\alias{print.bvsresponse}
\title{Prints a 'bvsresponse' object to the console.}
\usage{print.bvsresponse(x, ...)
}
\description{Prints a 'bvsresponse' object to the console.}
\value{Prints the 'bvsresponse' object to the console.}
\arguments{\item{x}{A 'bvsresponse' object}
\item{...}{unused}
}

