\name{print.parental}
\alias{print.parental}
\title{Prints a parental object to the console.}
\usage{\method{print}{parental} (x, ...)}
\description{Prints a \code{parental} object to the console.}
\value{Prints the \code{parental} object to the console.}
\arguments{\item{x}{A \code{parental} object}
\item{...}{Further arguments (unused)}}
