\name{print.parental}
\alias{print.parental}
\title{Prints 'parental'.}
\usage{\method{print}{parental} (x, ...)}

\description{
  Prints 'parental'.
}
\value{Prints the \code{parental} object to the console.}
\seealso{\code{\link{parental}}}
\arguments{
  \item{x}{A \code{parental} object}
  \item{...}{Further arguments, currently unused}
}
\examples{p1 <- parental(c(), c(1), c(1, 2, 4), c(3))
p1}
