\name{renameNodes.parental.list}
\alias{renameNodes.parental.list}
\title{Return the parental list, with nodes in each parental...}
\usage{\method{renameNodes}{parental.list} (x, newnames, ...)}
\description{Return the parental list, with nodes in each parental
in the parental.list renamed to the newnames.}
\value{The parental.list with renamed nodes}
\arguments{\item{x}{A parental.list}
\item{newnames}{A character vector specifying the new names for the 
nodes}
\item{...}{unused}}
