\name{toPolar}
\alias{toPolar}
\title{Convert cartesian coordinates to polars...}
\usage{toPolar(x, y)}
\description{Convert cartesian coordinates to polars}
\details{Currently unused. Not sure if this is vectorised}
\value{A list with two components, named \code{r} and \code{theta}, 
containing the radius and angle to the supplied point.}
\arguments{\item{x}{Cartesian x-coordinates}
\item{y}{Cartesian y-coordinates}}
