\name{topologicallyOrder}
\alias{topologicallyOrder}
\title{Topological ordering...}
\usage{topologicallyOrder(parental)}
\description{Topological ordering}
\details{Finds a permutation of the nodes 1, ..., p such that each all ancestors of 
each node have a lower place in the order.}
\value{A numeric vector. The order.}
\arguments{\item{parental}{An object of class \code{parental}}}
